% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lights.R
\name{hue_lights_state}
\alias{hue_lights_state}
\title{Set light state}
\usage{
hue_lights_state(id, ..., quiet = FALSE)
}
\arguments{
\item{id}{light id as numeric.}

\item{...}{Pass to body arguments.}

\item{quiet}{logical; suppress info on light state.}
}
\description{
Set light state
}
\seealso{
\url{https://developers.meethue.com/develop/hue-api/lights-api/}
}
