% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lights.R
\name{light_switch}
\alias{light_switch}
\title{Turn on/off on light a function later}
\usage{
light_switch(func, id, on = TRUE, quiet = FALSE, ..., delay = 0)
}
\arguments{
\item{func}{A function or formula (see \code{\link[rlang:as_function]{rlang::as_function()}}).}

\item{id}{light id as numeric.}

\item{on}{logical; Whether to turn on the lights or not.}

\item{quiet}{logical; suppress info on light state.}

\item{...}{Pass to body arguments.}

\item{delay}{Number of seconds in the future to delay execution. There is no
guarantee that the function will be executed at the desired time, but it
should not execute earlier.}
}
\description{
Turn on/off on light a function later
}
